examples/agent-runtime/agentic-frameworks/langgraph/requirements.txt
# Novita Agent Runtime
novita-sandbox>=1.1.0b1

# LangChain stack
langgraph>=0.6.7,<0.7.0
langchain-openai>=0.3.33,<0.4.0
langchain-community>=0.3.29,<0.4.0

# Tools and utilities
python-dotenv>=1.0.0,<=6.0.0
ddgs>=9.6.0,<10.0.0
